+++ js/widgets/behaviors/shapes/text.ts
@@ -27,27 +27,34 @@ export class TextShapeModel extends ShapeBaseModel {
     background: widget_serialization,
     padding: widget_serialization,
     scale_on_zoom: widget_serialization,
+    line_dash: widget_serialization,
   };
 
   drawNode2D(options: INodeCanvasBehaveOptions): void {
     const { context, node, globalScale } = options;
     const { x, y } = node;
 
-    this._drawCanvas({
+    const drawOptions = {
       ...TEXT_DEFAULTS,
       context,
       globalScale,
       x,
       y,
       ...this._resolveFacets(options),
-    });
+    };
+
+    if (drawOptions.text == null || !drawOptions.text.trim().length) {
+      return;
+    }
+
+    this._drawCanvas(drawOptions);
   }
 
-  drawNode3D(options: INodeThreeBehaveOptions): SpriteText {
+  drawNode3D(options: INodeThreeBehaveOptions): SpriteText | null {
     const { node, iframeClasses } = options;
     const { x, y } = node;
 
-    return this._drawThree({
+    const drawOptions = {
       ...TEXT_DEFAULTS,
       context: null,
       globalScale: null,
@@ -55,7 +62,13 @@ export class TextShapeModel extends ShapeBaseModel {
       y,
       iframeClasses,
       ...this._resolveFacets(options),
-    });
+    };
+
+    if (!drawOptions.text) {
+      return null;
+    }
+
+    return this._drawThree(drawOptions);
   }
 
   protected _drawThree(options: ITextOptions & IBaseOptions): SpriteText {
@@ -114,6 +127,7 @@ export class TextShapeModel extends ShapeBaseModel {
       scale_on_zoom,
       stroke_width,
       stroke,
+      line_dash,
     } = {
       ...TEXT_DEFAULTS,
       ...options,
@@ -132,6 +146,7 @@ export class TextShapeModel extends ShapeBaseModel {
     context.textBaseline = 'middle';
 
     if (stroke) {
+      context.setLineDash(line_dash || []);
       context.strokeStyle = stroke;
       context.lineWidth = scale_on_zoom ? stroke_width / globalScale : stroke_width;
       context.strokeText(text, x, y);